+++ src/Controller/Decision/DeleteDecisionController.php
@@ -12,11 +12,7 @@ use Symfony\Contracts\Translation\TranslatorInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
-/**
- * @IsGranted("ROLE_USER", message="Accès refusé. Espace reservé uniquement aux abonnés")
- *
- */
-
+#[IsGranted('ROLE_USER', message: 'Accès refusé. Connectez-vous')]
 #[Route("/decision")]
 class DeleteDecisionController extends AbstractController
 {
@@ -51,9 +47,14 @@ class DeleteDecisionController extends AbstractController
             return $this->redirectToRoute('home_mainMenu');
         }
 
-        $decision = $this->decisionRepository->findOneBySlug([
+        $decision = $this->decisionRepository->findOneBy([
             'slug' => $slug
         ]);
+
+        if (!$decision) 
+        {
+            return $this->redirectToRoute('page_error');
+        }
 
         $this->em->remove($decision);
         $this->em->flush();
